First batch of fixes 🎨 (to be approved)
