+++ src/components/home/oldStore.js
@@ -1,47 +1,39 @@
-import { View, Text, StyleSheet, Dimensions, TouchableOpacity, ScrollView, Image } from 'react-native'
+import { View, Text, StyleSheet, ImageBackground, Dimensions, TouchableOpacity, ScrollView, Image } from 'react-native'
 import { useTranslation } from 'react-i18next'
-import FontAwesome5Icon from 'react-native-vector-icons/FontAwesome5'
 
+import useGameContext from 'contexts/Game'
 import Back from 'components/basics/Back'
 import Coins from 'components/common/Coins'
 
-import LearningImage from 'assets/images/miscellaneous/learning.png'
+import ShopImage from 'assets/images/miscellaneous/shop.png'
 import { LinearGradient } from 'expo-linear-gradient'
 
 const { width, height } = Dimensions.get('window')
 const SHOP_IMAGE_RATIO = 599/416
 const SHOP_IMAGE_WIDTH = width * .6
-const STORE_OPTION_FONT_SIZE = height * .03
+const STORE_OPTION_FONT_SIZE = height * .05
 const SHOP_IMAGE_HEIGHT = SHOP_IMAGE_WIDTH / SHOP_IMAGE_RATIO
-const ICON_SIZE = width * .08
-const ICON_COLOR = 'white'
 
-const Learn = ({ navigation }) => {
+const Store = ({ navigation }) => {
 	const { t } = useTranslation()
 
   return (
-		<LinearGradient colors={['skyblue', 'lightblue']} style={{flex: 1}}>
+		<LinearGradient colors={['red', 'lightblue']} style={{flex: 1}}>
 			<Back onPress={() => navigation.navigate('Home')} color={'white'} />
 			<Coins containerStyle={styles.coinsContainer} />
 			<ScrollView contentContainerStyle={styles.container}>
 				<View style={styles.storeImageContainer}>
 					<Image
-						source={LearningImage}
+						source={ShopImage}
 						style={{width: SHOP_IMAGE_WIDTH, height: SHOP_IMAGE_HEIGHT }}
 					/>
 				</View>
 				<View style={styles.buttonsContainer}>
-					<TouchableOpacity onPress={() => navigation.navigate('World')} style={styles.buttonContainer}>
-						<FontAwesome5Icon name={"globe-europe"} size={ICON_SIZE} color={ICON_COLOR} />
-						<Text style={styles.optionStoreText}>{t("navigationTitle.world")}</Text>
+					<TouchableOpacity onPress={() => navigation.navigate('MapStyleStore')} style={styles.buttonContainer}>
+						<Text style={styles.optionStoreText}>{t("mapStyles")}</Text>
 					</TouchableOpacity>
-					<TouchableOpacity onPress={() => navigation.navigate('ExploreRegionPick')} style={styles.buttonContainer}>
-						<FontAwesome5Icon name={"map-marked-alt"} size={ICON_SIZE} color={ICON_COLOR} />
-						<Text style={styles.optionStoreText}>{t("navigationTitle.explore")}</Text>
-					</TouchableOpacity>
-          			<TouchableOpacity onPress={() => navigation.navigate('LearnRegionPick')} style={styles.buttonContainer}>
-						<FontAwesome5Icon name={"graduation-cap"} size={ICON_SIZE} color={ICON_COLOR} />
-						<Text style={styles.optionStoreText}>{t("navigationTitle.learn")}</Text>
+					<TouchableOpacity onPress={() => navigation.navigate('MarkerStyleStore')} style={styles.buttonContainer}>
+						<Text style={styles.optionStoreText}>{t("markerStyles")}</Text>
 					</TouchableOpacity>
 				</View>
 			</ScrollView>
@@ -71,8 +63,6 @@ const styles = StyleSheet.create({
 		fontSize: STORE_OPTION_FONT_SIZE,
 		fontFamily: 'Parisine',
 		textAlign: 'center',
-		marginTop: 5,
-		color: 'white'
 	},
 	buttonsContainer: {
 		flex: 1,
@@ -84,11 +74,8 @@ const styles = StyleSheet.create({
 		padding: width * .04,
 		alignItems: 'center',
 		marginVertical: height * .015,
-		backgroundColor: 'skyblue',
-		borderRadius: 20,
-		borderColor: 'white',
-		borderWidth: 1,
+		backgroundColor: '#CC5500	'
 	}
 })
 
-export default Learn
+export default Store
